feat(component): rename image
