+++ packages/react/src/components/Image/__stories__/index.scss
