+++ packages/react/src/components/Image/index.js
@@ -4,4 +4,4 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-export { default as ImageComponent } from './ImageComponent';
+export { default as Image } from './Image';
