+++ packages/react/src/components/index.js
@@ -18,4 +18,4 @@ export * from './Layout';
 export * from './ExpressiveModal';
 export * from './carbon-components-react/UIShell';
 export * from './TableOfContents';
-export * from './ImageComponent';
+export * from './Image';
